EraValidators moved to `primitives`
