+++ aleph-client/Cargo.toml
@@ -26,7 +26,6 @@ pallet-balances = { git = "https://github.com/Cardinal-Cryptography/substrate.gi
 pallet-vesting = { git = "https://github.com/Cardinal-Cryptography/substrate.git", branch = "aleph-v0.9.26", default-features = false }
 
 pallet-aleph = { path = "../pallets/aleph", default-features = false }
-pallet-elections = { path = "../pallets/elections", default-features = false }
 primitives = { path = "../primitives", default-features = false }
 
 [features]
