+++ pallets/elections/src/impls.rs
@@ -1,6 +1,6 @@
 use frame_election_provider_support::sp_arithmetic::Perquintill;
 use frame_support::{log::debug, pallet_prelude::Get};
-use primitives::CommitteeSeats;
+use primitives::{CommitteeSeats, EraValidators};
 use sp_staking::{EraIndex, SessionIndex};
 use sp_std::{
     collections::{btree_map::BTreeMap, btree_set::BTreeSet},
@@ -9,7 +9,7 @@ use sp_std::{
 
 use crate::{
     traits::{EraInfoProvider, SessionInfoProvider, ValidatorRewardsHandler},
-    CommitteeSize, Config, CurrentEraValidators, EraValidators, NextEraCommitteeSize,
+    CommitteeSize, Config, CurrentEraValidators, NextEraCommitteeSize,
     NextEraNonReservedValidators, NextEraReservedValidators, Pallet, SessionValidatorBlockCount,
     ValidatorEraTotalReward, ValidatorTotalRewards,
 };
